Command line arguments now actually work
* Convert IP to bytes representation
* Resolve MAC from IP
* Get MAC address
* Convert MAC to bytes representation
* Implemented rate command line arg

I have a few reservations on using system commands, but I think it works
for the purpose of prototyping. We are sending packets in such a way there
is little portability, so the system commands shouldn't be much worse.

The alternatives are to use a library (which would seem to give spotty results
anyway), to incorporate our sniffer with our attacker (lets not do that..) or
to force the user to provide everything to build the packet with (why would we
do that?). In the end, I think the solution used in this commit is reasonable.