Added -f, -b and ARP resolution
* Implemented -f (ip forwarding)
* Implemented -b (spoof both given IPs)
* Implemented ARP resolution before program start
* Updated status of implemented/unimplemented in comments

Note: I couldn't make a MITM attack with my setup, though
it is theoretically possible.

Besides the stuff mentioned in code, I'm trying to find
ways to ensure ip forwarding it working. Also, it would
be helpful to invoke arptables from the script like so:

    `arptables -A OUTPUT --source-ip 192.168.86.108 -j DROP`

The source-ip should be the ip of the computer that runs
the script. I am using this to avoid detection, however it
has limited use since a promiscuous sniffer would still
be able to detect a duplicate address. The DROP rule is
great for a simple attack that removes the victim's ability
to access the interneti though.